requirements.txt
# Core dependencies
pydantic>=2.0
python-dotenv>=1.0.0
requests>=2.31.0

# LLM Integration (choose one or more)
openai>=1.0.0
anthropic>=0.18.0
groq>=0.4.0

# LangGraph + LangChain (agentic pipeline)
langgraph>=0.2.0
langchain-core>=0.3.0
langchain-openai>=0.2.0
langchain-anthropic>=0.2.0
langchain-groq>=0.2.0

# Development
pytest>=7.4.0
black>=23.0.0
ruff>=0.1.0

# Audio processing
pydub>=0.25.1

# Image processing (Required for titles/captions)
Pillow>=10.0.0

# Database
pymongo>=4.6.0
dnspython>=2.4.0
